move navbar into own component folder and extract menu buttons into own component for more complex display and positioning logic
