+++ react-app/src/components/NavBar/NavBar.css
@@ -5,14 +5,12 @@
   align-items: center;
   position: sticky;
   top: 0;
-  padding-top: 3vh;
   z-index: 100;
   width: 100%;
   min-height: 145px;
   max-height: 145px;
   background-color: white;
   box-shadow: 0em 1em 1em rgba(155, 155, 155, 0.2);
-  padding-bottom: 2vh;
 }
 
 .div__transparent_bg {
@@ -34,17 +32,21 @@
 }
 
 .navbar_links {
+  z-index: 999;
+  display: flex;
+  flex-direction: column;
   align-items: center;
   position: absolute;
-  top: 10vh;
   background-color: white;
   border: 1px solid lightgrey;
-  padding: 20px;
+  padding: 5px 20px;
   border-radius: 20px;
+  width: 110px;
 }
 
 .div__user_button {
-  width: 3.5em;
+  width: 80px;
+  height: 55px;
   padding: 0.5em;
   border-radius: 18px;
   font-size: 1.4em;
@@ -54,22 +56,36 @@
   align-items: center;
 }
 
-.div__user_button div {
+.menu-icon {
   padding-top: 2px;
   color: #ff3a5c;
-  cursor: pointer;
 }
 
-.div__user_button button {
+.navbar_links > * {
+  margin: 5px 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.navbar_links:hover {
+  cursor: unset;
+}
+
+.navbar_links button {
   font-size: 15px;
   background-color: white;
   border: none;
 }
 
-.div__user_button button:focus {
+.navbar_links button:focus {
   outline: none;
 }
 
 .div__user_button:hover {
   cursor: pointer;
+}
+
+.button-usermenu {
+  padding: 0px;
 }
