--- react-app/src/components/auth/LogoutButton.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-
-import { LogOut } from '../../store/session';
-
-export default function LogoutButton () {
-  const dispatch = useDispatch();
-  const user = useSelector(state => state.session.user);
-
-  const onLogout = () => dispatch(LogOut());
-
-  return user
-    ? (
-      <button onClick={onLogout}>
-        Logout
-      </button>
-      )
-    : null;
-}
